move time to utils
