+++ utils/time.go
@@ -1,8 +1,8 @@
-package configuration
+package utils
 
 import "time"
 
 // Convert int (seconds) to [time.Duration]
-func intToSecond(numberOfSeconds int) time.Duration {
+func IntToSecond(numberOfSeconds int) time.Duration {
 	return time.Duration(numberOfSeconds) * time.Second
 }
