+++ utils/time_test.go
@@ -1,4 +1,4 @@
-package configuration
+package utils
 
 import (
 	"testing"
@@ -10,19 +10,19 @@ import (
 func TestIntToSecond(t *testing.T) {
 	assert.Equal(
 		t,
-		intToSecond(20),
+		IntToSecond(20),
 		time.Duration(20*time.Second),
 		"the duration should be equals",
 	)
 	assert.Equal(
 		t,
-		intToSecond(-5),
+		IntToSecond(-5),
 		time.Duration(-5*time.Second),
 		"the duration should be equals",
 	)
 	assert.Equal(
 		t,
-		intToSecond(3600),
+		IntToSecond(3600),
 		time.Duration(time.Hour),
 		"the duration should be equals",
 	)
